move ebanx specific code into more generic 'lib/checkout' because it's going to be used for things other than CC payments as well
