+++ client/components/credit-card-form-fields/test/index.js
@@ -14,14 +14,15 @@ import { identity, noop } from 'lodash';
  * Internal dependencies
  */
 import { CreditCardFormFields } from '../';
-import { shouldRenderAdditionalEbanxFields } from 'lib/credit-card-details/ebanx';
+import { shouldRenderAdditionalEbanxFields } from 'lib/checkout/ebanx';
+
 import mockCountriesList from './mocks/mock-countries-list';
 
 jest.mock( 'i18n-calypso', () => ( {
 	localize: x => x,
 } ) );
 
-jest.mock( 'lib/credit-card-details/ebanx', () => {
+jest.mock( 'lib/checkout/ebanx', () => {
 	return {
 		shouldRenderAdditionalEbanxFields: jest.fn( false ),
 	};
