+++ client/lib/checkout/constants.js
